Correct Tagged Heap And Arena Bookkeeping
Preserve full-width tag sequences, clear released arena slots
atomically, and
ensure heap allocation scans always release their lock.

Add comprehensive heap, arena, and job-system tests covering
concurrency,
lifetimes, alignment, ownership, statistics, and failure cases.

+++ include/ember/memory/tagged_heap.h
@@ -30,7 +30,7 @@ namespace ember
 	[[nodiscard]] constexpr u8 kind(HeapTag tag) noexcept { return static_cast<u8>(tag >> HEAP_TAG_SEQUENCE_BITS); }
 
 	/** Retrieves the `sequence` bits from a HeapTag */
-	[[nodiscard]] constexpr u8 sequence(HeapTag tag) noexcept { return tag & HEAP_TAG_SEQUENCE_MASK; }
+	[[nodiscard]] constexpr u64 sequence(HeapTag tag) noexcept { return tag & HEAP_TAG_SEQUENCE_MASK; }
 
 	/**
 	 * The shared block pool behind every allocator, following the tagged heap design in Christian
